Soft Launch
Home page is done for now (still need to mess with heros, but it's not as simple as I thought).

Menu and Footer are updated with Home,About,Postcards links and Footer is updated with Insta, Tumblr, and Email (email verified).

Some links on postcards page link to Etsy. Dead links are dead. Need to update Etsy. Need to create detail pages on haha (or just use Etsy...)